Refactor decoration unit tests
