+++ tests/unit/StatementTest.php
@@ -4,34 +4,14 @@ declare(strict_types=1);
 
 namespace Facile\DoctrineMySQLComeBack\Doctrine\DBAL;
 
-use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 
-class StatementTest extends TestCase
+class StatementTest extends AbstractDecoratorTestCase
 {
     use ProphecyTrait;
 
-    /**
-     * @dataProvider publicMethodsDataProvider
-     */
-    public function testAllParentMethodsAreDecorated(string $methodName): void
+    protected function getClassUnderTest(): string
     {
-        $connection = new \ReflectionClass(Statement::class);
-        $method = $connection->getMethod($methodName);
-
-        $this->assertEquals($connection, $method->getDeclaringClass(), 'Method not decorated: ' . $method->getName());
-    }
-
-    public function publicMethodsDataProvider(): \Generator
-    {
-        $dbalClass = new \ReflectionClass(\Doctrine\DBAL\Statement::class);
-
-        foreach ($dbalClass->getMethods() as $method) {
-            if (! $method->isPublic()) {
-                continue;
-            }
-
-            yield [$method->getName()];
-        }
+        return Statement::class;
     }
 }
